fix: :bug: Addition of literalMidWordUnderscores in showdown config to prevent sanitizedHTMLFromMarkdown from breaking
Testing the RenkuMarkdown component with $`\zeta_t \sim N(0,\Sigma_z(t))`$ caused the sanitizedHTMLFromMarkdown to break. Actually the inline markdown conversion of formulas was already working (i.e. testing it with $`E=mc^2`$ works). The error was that the underscore inside the formula was translated into <em>, breaking the conversion. Adding literalMidWordUnderscores: true in showDownOptions solves the issue. As side note, I've updated the showdown-katex and showdown packages and added the RenkuMarkdown.stories.tsx in order to conveniently test the component inside storybook.

1719, 2308