fix(client): fix the workflow view crashing the UI in some cases
Fixes #2578.

The "last execution" tooltip in the "Workflows" view can sometimes crash
the UI because the worflow id is not sanitized to create a valid CSS id.

This code change makes use of a `ref` instead of assigning a CSS id.

Also, the component has been refactored to make the rendering logic more
readable.