chore: update telepresence scripts (#2989)
Update the `telepresence` to be inline with the current version of [Telepresence OSS](https://www.getambassador.io/docs/telepresence-oss/latest/install).

The current version prefers that the `connect` command is used before the `intercept` command. The scripts now check the `telepresence status` and re-connect to the correct k8s namespace if needed.