refactor(client): transform pages into lazy components (#2984)
This refactor moves all pages imported at the root of the UI (from `App.jsx`) into lazy components. This change reduces the root bundle size and slightly improves the Lighthouse metrics. The main benefit of this change however is to improve the DevX around using `telepresence`: when reloading the UI after a code change, a much smaller number of source files will be reloaded by `vite` which makes development of the UI much smoother.

* Added `<LazyXYZPage />` components for each page defined in `App.jsx`
* Removed corresponding barrel files (see https://vitejs.dev/guide/performance.html#avoid-barrel-files)
* Moved around some components to fix build warnings